fix table link style and tabs refactor
