+++ packages/manager/src/components/TabLinkList/TabLinkList.tsx
@@ -2,53 +2,6 @@ import { Link } from '@reach/router';
 import * as React from 'react';
 import Tab from 'src/components/core/ReachTab';
 import TabList from 'src/components/core/ReachTabList';
-import { makeStyles, Theme } from 'src/components/core/styles';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  tab: {
-    '&[data-reach-tab]': {
-      display: 'inline-flex',
-      flexShrink: 0,
-      alignItems: 'center',
-      borderBottom: '2px solid transparent',
-      color: theme.textColors.linkActiveLight,
-      fontSize: '0.9rem',
-      lineHeight: 1.3,
-      marginTop: theme.spacing(0.5),
-      maxWidth: 264,
-      minHeight: theme.spacing(5),
-      minWidth: 50,
-      padding: '6px 16px',
-      textDecoration: 'none',
-      '&:focus': {
-        backgroundColor: theme.color.grey7,
-      },
-      '&:hover': {
-        backgroundColor: theme.color.grey7,
-        color: theme.palette.primary.main,
-      },
-    },
-    '&[data-reach-tab][data-selected]': {
-      borderBottom: `3px solid ${theme.textColors.linkActiveLight}`,
-      color: theme.textColors.headlineStatic,
-      fontFamily: theme.font.bold,
-    },
-  },
-  tabList: {
-    color: theme.color.tableHeaderText,
-    '&[data-reach-tab-list]': {
-      background: 'none !important',
-      boxShadow: `inset 0 -1px 0 ${
-        theme.name === 'lightTheme' ? '#e3e5e8' : '#2e3238'
-      }`,
-      marginBottom: theme.spacing(),
-      [theme.breakpoints.down('md')]: {
-        overflowX: 'auto',
-        padding: 1,
-      },
-    },
-  },
-}));
 
 export interface Tab {
   title: string;
@@ -61,23 +14,17 @@ interface Props {
   noLink?: boolean; // @todo: remove this prop if we use NavTab widely.
 }
 
-type CombinedProps = Props;
-
-export const TabLinkList: React.FC<CombinedProps> = (props) => {
-  const classes = useStyles();
-  const { tabs } = props;
-
+export const TabLinkList: React.FC<Props> = ({ tabs, noLink }) => {
   return (
-    <TabList className={classes.tabList}>
+    <TabList>
       {tabs.map((tab, _index) => {
         // @todo: remove this if we use NavTab widely.
-        const extraTemporaryProps: any = props.noLink
+        const extraTemporaryProps: any = noLink
           ? {}
           : { as: Link, to: tab.routeName };
 
         return (
           <Tab
-            className={classes.tab}
             key={`tab-${_index}`}
             // @todo: remove this if we use NavTab widely.
             {...extraTemporaryProps}
