--- packages/manager/src/components/core/ReachTab.ts
@@ -1,3 +1,0 @@
-import { Tab } from '@reach/tabs';
-
-export default Tab;
