--- packages/manager/src/components/core/ReachTabList.ts
@@ -1,3 +1,0 @@
-import { TabList } from '@reach/tabs';
-
-export default TabList;
